[project @ use stdint.h for 64-bit types]
Original author: Martin <martin@bzero.se>
Date: 2006-10-17 12:22:37+00:00

git-svn-id: http://shakespeer.googlecode.com/svn/trunk@101 3d55443d-7b4c-0410-a533-cbe148fed889

+++ gui/Aqua/NSStringExtensions.h
@@ -27,7 +27,7 @@
 - (NSString *)lastWindowsPathComponent;
 - (int)indexOfCharacter:(unichar)character fromIndex:(int)startIndex;
 - (int)occurencesOfString:(NSString *)aString;
-- (unsigned long long)unsignedLongLongValue;
+- (uint64_t)unsignedLongLongValue;
 - (NSAttributedString *)truncatedString:(NSLineBreakMode)mode;
 
 @end
